third_party/WebKit/LayoutTests/platform/fuchsia/external/wpt/css/cssom/medialist-interfaces-004-expected.txt

This is a testharness.js-based test.
PASS appendMedium_correctly_appends_medium_to_empty_MediaList
FAIL appendMedium_correctly_appends_medium_to_nonempty_MediaList assert_equals: expected "all, screen" but got "screen, all"
Harness: the test ran to completion.
